Relación 1 a N con todas las funcionalidades protegidas

+++ app/Http/Controllers/NoteController.php
@@ -4,17 +4,24 @@ namespace App\Http\Controllers;
 
 use App\Models\Note;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class NoteController extends Controller
 {
     public function store(Request $request)
-    {
-       
+    {   
+        if (!Auth::user()) {
+        return redirect()->back();
+        }
         $validated = $request->validate([
-            'title' => 'required|max:100',
-            'content' => 'required',
+        'title' => 'required|max:100',
+        'content' => 'required',
         ]);
-        Note::create($validated);
+        
+        $user=Auth::user(); 
+        $user->notes()->create($validated);
+
+        /* Note::create($validated); */
         /* La misma forma sin simplificar */
         /* Note::create([
             'title' => $validated['title'],
@@ -33,8 +40,17 @@ class NoteController extends Controller
 
     public function update($id, Request $request)
     {
+        if (!Auth::user()) {
+            return redirect()->back();
+            }
+
+        
         //Para modificar algo, hay que recuperar lo que quiero modificar
         $note = Note::findOrFail($id);
+
+        if (Auth::id()!=$note->user_id) {
+            return redirect()->back();
+            }
 
         //Antes de modificar, recupero los datos con los cuales quiero modificar los datos
         $validateddata = $request->validate([
@@ -55,8 +71,17 @@ class NoteController extends Controller
 
     public function delete($id)
     {   
+        if (!Auth::user()) {
+            return redirect()->back();
+            }
+    
+
         //Recuperamos el dato que queremos borrar
         $note = Note::findOrFail($id);
+
+        if (Auth::id()!=$note->user_id) {
+            return redirect()->back();
+            }
 
         //Borramos la nota y salimos
         $note->delete();
